Add Document-based example Add NSString versions of all formats and master call to allow display Break out MILFormats to allow use of format selector without inheriting all of MediaInfo
