+++ MediaInfoDemo/MediaInfoDocument/AppDelegate.h
@@ -1,0 +1,15 @@
+//
+//  AppDelegate.h
+//  MediaInfo Document
+//
+//  Created by Hugh Mackworth on 2/18/16.
+//  Copyright © 2016 Hugh Mackworth. All rights reserved.
+//
+
+#import <Cocoa/Cocoa.h>
+
+@interface AppDelegate : NSObject <NSApplicationDelegate>
+
+
+@end
+
