+++ MediaInfoDemo/MediaInfoDocument/MIKViewController.h
@@ -1,0 +1,14 @@
+//
+//  MIKViewController.h
+//  MediaInfo Document
+//
+//  Created by Hugh Mackworth on 2/18/16.
+//  Copyright © 2016 Hugh Mackworth. All rights reserved.
+//
+
+#import <Cocoa/Cocoa.h>
+
+@interface MIKViewController : NSViewController
+
+@end
+
