+++ MediaInfoDemo/MediaInfoDocument/main.m
@@ -1,0 +1,13 @@
+//
+//  main.m
+//  MediaInfo Document
+//
+//  Created by Hugh Mackworth on 2/18/16.
+//  Copyright © 2016 Hugh Mackworth. All rights reserved.
+//
+
+#import <Cocoa/Cocoa.h>
+
+int main(int argc, const char * argv[]) {
+    return NSApplicationMain(argc, argv);
+}
